Vislib image test
