+++ tests/test_graphics.c
@@ -5,10 +5,6 @@
 #include "proto.h"
 #include "tests.h"
 
-BEGIN_INCL_NO_WARN
-#include "../external/stb_image.h"
-END_INCL_NO_WARN
-
 
 
 /*************************************************************************************************/
@@ -227,58 +223,6 @@ static int _graphics_screenshot(TestGraphics* tg, const char* name)
     snprintf(path, sizeof(path), "%s/docs/images/graphics/%s.png", ROOT_DIR, name);
     dvz_screenshot_file(tg->canvas, path);
     return res;
-}
-
-
-
-static DvzTexture* _earth_texture(DvzContext* context)
-{
-    DvzGpu* gpu = context->gpu;
-    char path[1024];
-    snprintf(path, sizeof(path), "%s/textures/earth.jpg", DATA_DIR);
-    int width, height, depth;
-    uint8_t* tex_data = stbi_load(path, &width, &height, &depth, STBI_rgb_alpha);
-    uint32_t tex_size = (uint32_t)(width * height);
-    DvzTexture* texture = dvz_ctx_texture(
-        gpu->context, 2, (uvec3){(uint32_t)width, (uint32_t)height, 1}, VK_FORMAT_R8G8B8A8_UNORM);
-    dvz_upload_texture(
-        context, texture, DVZ_ZERO_OFFSET, DVZ_ZERO_OFFSET, tex_size * sizeof(cvec4), tex_data);
-    FREE(tex_data)
-    return texture;
-}
-
-static DvzTexture* _volume_texture(DvzContext* context, int kind)
-{
-    const uint32_t S = 64;
-    VkDeviceSize size = S * S * S * sizeof(uint8_t);
-    DvzTexture* texture = dvz_ctx_texture(context, 3, (uvec3){S, S, S}, VK_FORMAT_R8_UNORM);
-    uint8_t* tex_data = calloc(S * S * S, sizeof(uint8_t));
-    uint32_t l = 0;
-    double x, y, z, w;
-    double c = S / 2;
-    for (uint32_t i = 0; i < S; i++)
-    {
-        for (uint32_t j = 0; j < S; j++)
-        {
-            for (uint32_t k = 0; k < S; k++)
-            {
-                x = ((double)i - c) / c;
-                y = ((double)j - c) / c;
-                z = ((double)k - c) / c;
-                w = exp(-4 * (x * x + y * y + z * z));
-
-                if (kind == 0)
-                    tex_data[l++] = TO_BYTE(w);
-                else
-                    tex_data[l++] = dvz_rand_byte() % 3;
-
-                // tex_data[l++] = (i & j) | (i & k) | (j & k) ? 0 : 32;
-            }
-        }
-    }
-    dvz_upload_texture(context, texture, DVZ_ZERO_OFFSET, DVZ_ZERO_OFFSET, size, tex_data);
-    FREE(tex_data);
-    return texture;
 }
