+++ tests/tests.h
@@ -134,7 +134,7 @@ int test_vislib_triangle_fan(TestContext*);
 int test_vislib_marker(TestContext*);
 int test_vislib_polygon(TestContext*);
 int test_vislib_path(TestContext*);
-int test_vislib_image(TestContext*);
+int test_vislib_image_1(TestContext*);
 int test_vislib_image_cmap(TestContext*);
 int test_vislib_axes(TestContext*);
 int test_vislib_mesh(TestContext*);
@@ -273,7 +273,7 @@ static TestCase TEST_CASES[] = {
     CASE_FIXTURE(CANVAS, test_vislib_marker),         //
     CASE_FIXTURE(CANVAS, test_vislib_polygon),        //
     CASE_FIXTURE(CANVAS, test_vislib_path),           //
-    CASE_FIXTURE(CANVAS, test_vislib_image),          //
+    CASE_FIXTURE(CANVAS, test_vislib_image_1),        //
     CASE_FIXTURE(CANVAS, test_vislib_image_cmap),     //
     CASE_FIXTURE(CANVAS, test_vislib_axes),           //
     CASE_FIXTURE(CANVAS, test_vislib_mesh),           //
